moved source code to src dir
