+++ src/auth.h
